Colocate workflow with test
