+++ packages/test/src/test-integration-update-signal-query-example.ts
@@ -1,14 +1,8 @@
 import * as wf from '@temporalio/workflow';
 import { helpers, makeTestFunction } from './helpers-integration';
-import {
-  counterWorkflow,
-  getValueQuery,
-  incrementAndGetValueUpdate,
-  incrementSignal,
-} from './workflows/update-signal-query-example';
 
 const test = makeTestFunction({
-  workflowsPath: require.resolve('./workflows/update-signal-query-example'),
+  workflowsPath: __filename,
   workflowEnvironmentOpts: {
     // TODO: remove this server config when default test server supports update
     server: {
@@ -19,6 +13,27 @@ const test = makeTestFunction({
     },
   },
 });
+
+/* Example from WorkflowHandle docstring */
+
+// @@@SNIPSTART typescript-workflow-update-signal-query-example
+export const incrementSignal = wf.defineSignal<[number]>('increment');
+export const getValueQuery = wf.defineQuery<number>('getValue');
+export const incrementAndGetValueUpdate = wf.defineUpdate<number, [number]>('incrementAndGetValue');
+
+export async function counterWorkflow(initialValue: number): Promise<void> {
+  let count = initialValue;
+  wf.setHandler(incrementSignal, (arg: number) => {
+    count += arg;
+  });
+  wf.setHandler(getValueQuery, () => count);
+  wf.setHandler(incrementAndGetValueUpdate, (arg: number): number => {
+    count += arg;
+    return count;
+  });
+  await wf.condition(() => false);
+}
+// @@@SNIPEND
 
 /* Example from WorkflowHandle docstring */
 test('Update/Signal/Query example in WorkflowHandle docstrings works', async (t) => {
