--- packages/test/src/workflows/update-signal-query-example.ts
@@ -1,22 +1,0 @@
-import * as wf from '@temporalio/workflow';
-
-/* Example from WorkflowHandle docstring */
-
-// @@@SNIPSTART typescript-workflow-update-signal-query-example
-export const incrementSignal = wf.defineSignal<[number]>('increment');
-export const getValueQuery = wf.defineQuery<number>('getValue');
-export const incrementAndGetValueUpdate = wf.defineUpdate<number, [number]>('incrementAndGetValue');
-
-export async function counterWorkflow(initialValue: number): Promise<void> {
-  let count = initialValue;
-  wf.setHandler(incrementSignal, (arg: number) => {
-    count += arg;
-  });
-  wf.setHandler(getValueQuery, () => count);
-  wf.setHandler(incrementAndGetValueUpdate, (arg: number): number => {
-    count += arg;
-    return count;
-  });
-  await wf.condition(() => false);
-}
-// @@@SNIPEND
